requirements.txt
pymongo==4.4.0
pyyaml==6.0
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
python-dateutil==2.8.2
rapidfuzz==2.14.0
